Change Int32 to Int64
Change the following properties from `Int32` to to `Int64` to avoid
`System.OverflowException` on `int.Parse()` of very large numbers:

- `SemanticVersion.Major`
- `SemanticVersion.Minor`
- `SemanticVersion.Patch`
- `SemanticVersionBuildMetaData.CommitsSinceTag`
- `SemanticVersionBuildMetaData.CommitsSinceVersionSource`
- `SemanticVersionBuildMetaData.UncommittedChanges`
- `SemanticVersionPreReleaseTag.Number`

+++ src/GitVersion.Core/VersionCalculation/SemanticVersioning/SemanticVersionBuildMetaData.cs
@@ -14,15 +14,15 @@ public class SemanticVersionBuildMetaData : IFormattable, IEquatable<SemanticVer
     private static readonly LambdaEqualityHelper<SemanticVersionBuildMetaData> EqualityHelper =
         new(x => x.CommitsSinceTag, x => x.Branch, x => x.Sha);
 
-    public int? CommitsSinceTag;
+    public long? CommitsSinceTag;
     public string? Branch;
     public string? Sha;
     public string? ShortSha;
     public string? OtherMetaData;
     public DateTimeOffset? CommitDate;
     public string? VersionSourceSha;
-    public int? CommitsSinceVersionSource;
-    public int UncommittedChanges;
+    public long? CommitsSinceVersionSource;
+    public long UncommittedChanges;
 
     public SemanticVersionBuildMetaData()
     {
@@ -124,7 +124,7 @@ public class SemanticVersionBuildMetaData : IFormattable, IEquatable<SemanticVer
 
         if (parsed.Groups["BuildNumber"].Success)
         {
-            semanticVersionBuildMetaData.CommitsSinceTag = int.Parse(parsed.Groups["BuildNumber"].Value);
+            semanticVersionBuildMetaData.CommitsSinceTag = long.Parse(parsed.Groups["BuildNumber"].Value);
             semanticVersionBuildMetaData.CommitsSinceVersionSource = semanticVersionBuildMetaData.CommitsSinceTag ?? 0;
         }
